requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.3
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brotli==1.0.7
brotlipy==0.7.0
cached-property==1.5.1
cachetools==5.3.0
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.34
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.7.3
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==6.0.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
empy==3.3.2
entrypoints==0.3
evdev==1.3.0
-e git+https://gitlab.com/acrome-colab/riders-poc/f1tenth-riders-quickstart@3f96c382a2b314ff775284ef2c2a8e278ceaa925#egg=f110_gym&subdirectory=gym
filelock==3.0.12
Flask==2.2.3
fonttools==4.37.1
fuse-python==1.0.0
future==0.18.2
genmsg==0.5.14
genpy==0.6.9
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gpg===1.16.0-unknown
greenlet==0.4.15
gssapi==1.6.1
gym==0.18.0
h11==0.13.0
html5lib==1.0.1
httplib2==0.21.0
idna==2.8
ifaddr==0.1.6
importlib-metadata==6.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.39.1
lxml==4.5.0
lz4==3.0.2+dfsg
MarkupSafe==2.1.2
matplotlib==3.5.3
more-itertools==4.2.0
msgpack==0.6.2
multidict==5.2.0
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
netifaces==0.10.4
nose==1.3.7
Nuitka==1.5.7
numba==0.56.2
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
ordered-set==4.1.0
outcome==1.1.0
packaging==21.3
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.11
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
py-cpuinfo==5.0.0
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pyglet==1.5.0
Pygments==2.3.1
PyGObject==3.36.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyJWT==1.7.1
pykerberos==1.1.14
pylibacl==0.5.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
PyOpenGL==3.1.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.15.5
PySocks==1.7.1
pytest==7.1.3
pytest-rerunfailures==10.2
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-gnupg==0.4.5
python-lzo==1.12
python-magic==0.4.16
pytz==2022.7.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
rencode==1.0.6
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rocker==0.2.10
roman==2.0.0
rosdep==0.22.1
rosdep-modules==0.22.1
rosdistro-modules==0.9.0
rospkg==1.4.0
rospkg-modules==1.4.0
rsa==4.9
scipy==1.9.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.1.3
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
ssh-import-id==5.10
testresources==2.0.1
texttable==1.6.2
tomli==2.0.1
tornado==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.0.1
ubuntu-drivers-common==0.0.0
ufw==0.36
uritemplate==4.1.1
uritools==3.0.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
virtualenv==20.0.17
wadllib==1.3.3
webdriver-manager==3.5.4
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.3
wsproto==1.1.0
xkit==0.0.0
xpra==3.0.6
yarg==0.1.9
yarl==1.7.2
zeroconf==0.24.4
zipp==1.0.0
zstandard==0.21.0
